explorer/claude/requirements.txt
google-generativeai>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
argparse>=1.4.0